neuroswipe/__init__.py
import pytorch_lightning as pl
import torch
import torchmetrics
from nip import nip

from neuroswipe import dataloaders, models, training

nip(dataloaders)
nip(models)
nip(training)

# torch
nip(torch.optim)
nip(torch.optim.lr_scheduler)
nip(torch.nn)
nip(torchmetrics)

# pytorch lightning
nip(pl.callbacks)
