neuroswipe/requirements.txt
torch==1.13.0
pytorch-lightning==1.7.0
rich==12.6.0
nip-config==0.8.3
ruamel-yaml==0.16.12
tqdm==4.66.1
fasttext==0.9.2
jsonlines==3.1.0
notebook==6.4.12
pandas==1.3.5